+++ Assign6.h
@@ -19,10 +19,10 @@ bool GetYesOrNoInput(const string initialMessage);
 int GetMenuInput(const string initialMessage, const int numMenuOptions);
 int GetPositiveIntFromConsole(string itemsToGetNumberOf);
 
-void GenerateRandomValuesForArray(float *&array, int numberOfValues);
-void CreateArrayFromFile(float *&array, int &arraySize, string fileName);
-void PrintDynamicArray(string arrayName, float *&array, int arraySize);
-void CopyArray(float *&origionalArray, int arraySize, float *&duplicateArray);
+void GenerateRandomValuesForArray(double *&array, int numberOfValues);
+void CreateArrayFromFile(double *&array, int &arraySize, string fileName);
+void PrintDynamicArray(string arrayName, double *&array, int arraySize);
+void CopyArray(double *&origionalArray, int arraySize, double *&duplicateArray);
 double CalcTotalTime(double startTime, double endTime);
 void UpdateFastestAndSlowest(double newTime, double &fastestTime, double &slowestTime,
                 string newSortName, string &fastestSortName, string &slowestSortName);
